Refactored comments tests
