+++ app/__init__.py
@@ -12,7 +12,7 @@ DATA_RECIPIENT = os.getenv('DATA_RECIPIENT', 'dap@ons.gov.uk')
 app = Flask(__name__)
 socketio = SocketIO(app)
 
-from app.datastore import DATASTORE_CLIENT, DATASTORE_NAMESPACE
+from app.datastore import DATASTORE_TOOLBOX_CLIENT, DATASTORE_TOOLBOX_NAMESPACE
 from app.messaging import get_message_manager
 message_manager = get_message_manager(listen)
